LinphoneActivity launcher (synchro service ready)
Due to a bug in tabhost on 1.5 devices (G1 and simulator)
it was impossible to have the synchronisation code directly
in LinphoneActivity. (NPE on touchmodechanged on mCurrentView).

Note that since the synchronisation code rewrite, no special care
is taken after a service crash. As a consequence you should always
check the root cause of a
"Caused by: java.lang.RuntimeException: Linphone Manager should be created before accessed"