Update speex_AndroidConfig.h
`#undef` is semantically somewhat of a typo here. 
Exactly 1 of: {FIXED_POINT, FLOATING_POINT} should be defined and set to 1.
In the speex_AndroidConfig.h file, there is a line `USE_FLOAT 0`. Changing this to 1 does not build,
because nowhere is:

`#define FLOAT_POINT 1` 

ever set. Truly, this is all in terrible error, as FLOATING_POINT and FIXED_POINT are mutually exclusive, there
should be a single flag controlling both. The current setup leads to states w/undefined behavior, and I am
even able to build libraries that cannot register accounts simply because these flags are poorly managed.