Support dialing with hw numpad keys
Combined with #2494 and with this setting:

```
[ui]
automatically_show_dialpad=1
```

The user can do basic calls with an hardware keyboard.

This also fixes an show/hide dialpad inconsistency, where the android IME and dialpad is shown at the same time when used in combination with #2494, or android IME is not hidden and dialpad is not shown automatically.

Basically it just hides the android IME before showing the dialpad with an delay and in the right threads.

To reproduce the first behaviour you have to use #2494, to reproduce the second one, do the following:
- Set `automatically_show_dialpad=1`
- open linphone, goto new call and press search field to open android IME
- Press home button on phone/tablet
- press recent button on phone/tablet
- select linphone
- Android IME will be shown